Finish utils module
Added chat hooks inside utils module.

Module devs are not expected to add chat hooks inside their
modules, but are recommended to do so for the sake of the bot devs.
I'll clean up chat hooks in later revisions so they can be more
easily overridden. For now, I just need to get this done.