Auto-remove song limits, fix bot autoDJ
Bot will remove song limits if there are not enough people on decks
and/or queue to warrant a limit. He will also block limits from
being imposed in the same conditions.

The bot also only ever checked for the hop-down condition at the
end of the song. Added a check on add_dj, but only if he's not
playing.