Change safeguard
Apparently TTAPI doesn't update the room data until the newsong event, so
the safeguard code was moved there. Also, the list will only ever be
replaced if the room data version is smaller. While this may not be
correct 100% of the time, it will prevent the bot from trying to remove
DJs incorrectly, which is more important.